refactor: post.goをmain.goに統合
